Fixes #1 and #4 - rotations and action bar items.
Figured out that all I needed to do to implement the fragment state
saving was to implement onSaveInstanceState and onCreate appropriately
in the fragments, and set up the onCreate of the activity to only attach
the list fragment if sis is null. Everything else, including the
attached fragment and the stack, is saved. So, good.

Found out how to have different menus show up for different fragments -
basically the onCreateOptionsMenu, plus a setHasOptionsMenu. Got rid of
the edit, launch, save, cancel buttons, and replaced 'em with action
items. also, set up home-as-up button, so there isn't even a cancel
button. decided not to use a contextual action bar - too complicated.