Fixes #5. Lots of changes to make it work.
The whole access thing is now a bit of a mess; deletes were not at all
straightforward. It's clear there are better architectures for
representing the set of stored items. I don't really feel like modifying
the whole thing again, though.

In order to make the profile list update correctly after deletion, have
to clear and re-fill the adapter. In order to make that work from the
dialog in the CAM, I used publisher/subscriber to allow the
ProfileDelete trait to present an onDeleteYes method. Better than
things I tried before that, though.